WIP: render_with_dimensions
In order to support scrollable viewports, the renderer can now keep
track of the dimensions for each element and child when rendering a box.
These are reported when using `render_with_dimensions`. This returns the
height and content_height of each element, which can be used to
calculate the offset required for scrolling.

TODO: get grids working again
TODO: make overflow-scroll a thing
TODO: add a scrollbar?